{fuse, sched/wait}: Replicate pre-TASK_FREEZABLE behavior on downstream
msm-5.4 lacks TASK_FREEZABLE which was introduced with freezer rewrite in
upstream [1]. That's a huge backport that's not worth for the small change in
fuse. Instead, replicate how the logic worked before the introduction of
TASK_FREEZABLE in upstream and adapt for our kernel.

Now wait_event_state_exclusive() is only exclusive for freezable tasks. So,
let's edit the macro name to wait_event_freezable_state_exclusive().

[1] https://github.com/torvalds/linux/commit/f5d39b020809146

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>